refactor: rename dp_solver -> bellman-solver
